# docs/swagger/swagger.yaml
basePath: /v0
consumes:
- application/json
definitions:
  github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse:
    properties:
      error:
        type: string
      namespace:
        type: string
      success:
        type: boolean
    type: object
  github_com_rabbytesoftware_quiver_core_internal_api_libs.MutationResponse:
    properties:
      namespace:
        type: string
      success:
        type: boolean
    type: object
  github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ArrowDependentsDTO:
    properties:
      dependents:
        items:
          type: string
        type: array
      namespace:
        type: string
    type: object
  github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ArrowDetailDTO:
    properties:
      active_run:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.RunRecordDTO'
      channel:
        type: string
      description:
        type: string
      inference:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.InferenceDTO'
      installed_at:
        type: string
      installed_constraint:
        type: string
      last_return:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ReturnDTO'
      last_used_at:
        type: string
      license:
        type: string
      name:
        type: string
      namespace:
        type: string
      origin:
        type: string
      outdated:
        type: boolean
      recommended_ref:
        type: string
      state:
        type: string
      tags:
        items:
          type: string
        type: array
      user_installed:
        type: boolean
    type: object
  github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ArrowListItemDTO:
    properties:
      confidence:
        type: string
      description:
        type: string
      media:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_domain.ArrowMedia'
      name:
        type: string
      namespace:
        type: string
      origin:
        type: string
      tags:
        items:
          type: string
        type: array
      versions:
        items:
          $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.InstalledVersionItemDTO'
        type: array
    type: object
...
  github_com_rabbytesoftware_quiver_core_internal_app_models.AddOptions:
    properties:
      channel:
        type: string
      confirm:
        type: boolean
    type: object
  github_com_rabbytesoftware_quiver_core_internal_app_models.UpdateOptions:
    properties:
      channel:
        type: string
      ref:
        type: string
      upgradeRef:
        type: boolean
    type: object
...
host: localhost:40257
info:
  contact:
    name: Rabbyte Software
  description: Quiver is a decentralized package manager.
  title: Quiver API
  version: "0"
paths:
  /arrow:
    get:
      description: Returns all registered arrows. Use the WebSocket upgrade to stream
        real-time updates instead.
      parameters:
      - description: Filter by user-installed flag
        in: query
        name: user_installed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ArrowListItemDTO'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: List arrows
      tags:
      - arrows
  /arrow/{ns}:
    delete:
      description: Deregisters an arrow, addressed by the namespace it was registered
        under (bare or versioned).
      parameters:
      - description: Arrow namespace (bare or versioned, e.g. github.com/user/repo@v1.0.0)
        in: path
        name: ns
        required: true
        type: string
      responses:
        "200":
          description: Arrow removed
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.MutationResponse'
        "404":
          description: Arrow not found
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Remove arrow
      tags:
      - arrows
    get:
      description: Returns detailed information for an arrow including current state,
        active run, and last return. An uncatalogued namespace resolves live, the
        same way the manifest and readme endpoints do, and reports state "absent".
        Use the WebSocket upgrade to stream live updates.
      parameters:
      - description: Arrow namespace
        in: path
        name: ns
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ArrowDetailDTO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Get arrow detail
      tags:
      - arrows
    patch:
      consumes:
      - application/json
      description: 'Fetches the latest manifest for the arrow and updates its registration.
        Optional body fields: "channel" switches which release channel the arrow tracks
        (its latest ref is taken unless "ref" pins to a specific ref within that channel);
        "upgrade_ref" resolves the arrow''s existing installed constraint to its latest
        matching ref instead.'
      parameters:
      - description: Arrow namespace
        in: path
        name: ns
        required: true
        type: string
      - description: Optional update preferences (e.g. channel, ref, upgrade_ref)
        in: body
        name: body
        schema:
          $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_app_models.UpdateOptions'
      responses:
        "200":
          description: Arrow updated
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.MutationResponse'
        "400":
          description: Requested channel or ref does not exist
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "404":
          description: Arrow not found
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Update arrow manifest
      tags:
      - arrows
    post:
      description: Registers an arrow by its namespace. The manifest must already
        exist in the registry. A low-confidence inferred arrow requires confirmation,
        either "confirm":true in the body or ?confirm=true.
      parameters:
      - description: Arrow namespace (e.g. github.com/user/repo@v1.0.0)
        in: path
        name: ns
        required: true
        type: string
      - description: Confirms adding a low-confidence inferred arrow
        in: query
        name: confirm
        type: boolean
      - description: Optional install preferences (e.g. channel, confirm)
        in: body
        name: body
        schema:
          $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_app_models.AddOptions'
      responses:
        "201":
          description: Arrow registered
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.MutationResponse'
        "400":
          description: Invalid namespace
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "404":
          description: Manifest not found
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "409":
          description: Arrow already registered, or confirmation required for a low-confidence
            inferred arrow
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Register arrow
      tags:
      - arrows
  /arrow/{ns}/channels:
    get:
      description: Buckets every tag the arrow's repository publishes into its release
        channel, plus the repository's default branch.
      parameters:
      - description: Arrow namespace
        in: path
        name: ns
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Channels
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ChannelListDTO'
              type: object
        "404":
          description: Arrow not found
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: List arrow channels
      tags:
      - arrows
  /arrow/{ns}/dependencies:
    get:
      description: Returns the transitive dependency plan for this arrow, resolved
        from its manifest's tools and services, in dependency-first order.
      parameters:
      - description: Arrow namespace
        in: path
        name: ns
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ArrowDependenciesDTO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Get arrow dependencies
      tags:
      - arrows
  /arrow/{ns}/dependents:
    get:
      description: Returns the namespaces of arrows that declare a dependency on this
        arrow, across all installed versions pointing at it.
      parameters:
      - description: Arrow namespace
        in: path
        name: ns
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ArrowDependentsDTO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Get arrow dependents
      tags:
      - arrows
  /arrow/{ns}/manifest:
    get:
      description: Returns the full manifest definition including targets, variables,
        and lifecycle steps.
      parameters:
      - description: Arrow namespace
        in: path
        name: ns
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ArrowManifestDTO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Get arrow manifest
      tags:
      - arrows
    post:
      consumes:
      - application/x-yaml
      description: Accepts a raw YAML manifest in the request body, stores it in the
        registry, and registers the arrow. Use this to publish a new arrow from a
        local manifest file.
      parameters:
      - description: Arrow namespace
        in: path
        name: ns
        required: true
        type: string
      - description: Raw YAML manifest
        in: body
        name: body
        required: true
        schema:
          type: string
      responses:
        "201":
          description: Manifest seeded and arrow registered
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.MutationResponse'
        "400":
          description: Failed to read body
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "422":
          description: Invalid manifest
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Seed arrow manifest
      tags:
      - arrows
  /arrow/{ns}/manifest/validate:
    post:
      consumes:
      - application/x-yaml
      description: Parses and validates a raw YAML manifest. Returns validation errors
        and platform compatibility without writing anything to the registry.
      parameters:
      - description: Arrow namespace
        in: path
        name: ns
        required: true
        type: string
      - description: Raw YAML manifest
        in: body
        name: body
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Manifest is valid
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ValidationResultDTO'
              type: object
        "400":
          description: Failed to read body
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "422":
          description: Manifest has validation errors
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ValidationResultDTO'
              type: object
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Validate arrow manifest
      tags:
      - arrows
  /arrow/{ns}/preview:
    get:
      description: Resolves an arrow's manifest without registering it, the same way
        GetDetail resolves an uncatalogued namespace live. ?format=raw returns the
        manifest bytes instead of the parsed detail.
      parameters:
      - description: Arrow namespace
        in: path
        name: ns
        required: true
        type: string
      - description: raw returns the manifest bytes as text/markdown
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ArrowDetailDTO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Preview arrow
      tags:
      - arrows
...
produces:
- application/json
swagger: "2.0"