# internal/api/v0/endpoints/arrows/handlers/handlers.go
package arrows

import (
	"io"
	"net/http"

	"github.com/gin-gonic/gin"

	"github.com/rabbytesoftware/quiver.core/internal/api/libs"
	"github.com/rabbytesoftware/quiver.core/internal/api/libs/apierr"
	apidto "github.com/rabbytesoftware/quiver.core/internal/api/v0/dto"
	"github.com/rabbytesoftware/quiver.core/internal/app/models"
	"github.com/rabbytesoftware/quiver.core/internal/app/usecases"
	"github.com/rabbytesoftware/quiver.core/internal/domain"
)

type Handlers struct {
	svc usecases.ArrowUsecase
}

func New(svc usecases.ArrowUsecase) *Handlers {
	return &Handlers{svc: svc}
}

// Add registers an arrow from an existing manifest in the Quiver registry.
//
// @Summary      Register arrow
// @Description  Registers an arrow by its namespace. The manifest must already exist in the registry. A low-confidence inferred arrow requires confirmation, either "confirm":true in the body or ?confirm=true.
// @Tags         arrows
// @Param        ns       path   string             true   "Arrow namespace (e.g. github.com/user/repo@v1.0.0)"
// @Param        confirm  query  bool               false  "Confirms adding a low-confidence inferred arrow"
// @Param        body     body   models.AddOptions  false  "Optional install preferences (e.g. channel, confirm)"
// @Success      201  {object}  libs.MutationResponse  "Arrow registered"
// @Failure      400  {object}  libs.ErrResponse       "Invalid namespace"
// @Failure      404  {object}  libs.ErrResponse       "Manifest not found"
// @Failure      409  {object}  libs.ErrResponse       "Arrow already registered, or confirmation required for a low-confidence inferred arrow"
// @Failure      500  {object}  libs.ErrResponse       "Internal error"
// @Router       /arrow/{ns} [post]
func (h *Handlers) Add(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))
	opts := models.AddOptions{}
	if c.Request.Body != nil {
		_ = c.ShouldBindJSON(&opts)
	}
	if c.Query("confirm") == "true" {
		opts.Confirm = true
	}
	if err := h.svc.Add(c.Request.Context(), ns, opts); err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns), err)
		return
	}
	libs.WriteMutationOK(c, http.StatusCreated, string(ns))
}

// Update pulls the latest manifest for an arrow from the registry and re-registers it.
// It also accepts an optional channel switch: setting channel moves the
// arrow onto a different release channel (taking that channel's latest ref
// unless ref pins to a specific member of it), independent of upgrade_ref's
// existing constraint-based upgrade.
//
// @Summary      Update arrow manifest
// @Description  Fetches the latest manifest for the arrow and updates its registration. Optional body fields: "channel" switches which release channel the arrow tracks (its latest ref is taken unless "ref" pins to a specific ref within that channel); "upgrade_ref" resolves the arrow's existing installed constraint to its latest matching ref instead.
// @Tags         arrows
// @Accept       json
// @Param        ns    path  string              true   "Arrow namespace"
// @Param        body  body  models.UpdateOptions  false  "Optional update preferences (e.g. channel, ref, upgrade_ref)"
// @Success      200  {object}  libs.MutationResponse  "Arrow updated"
// @Failure      400  {object}  libs.ErrResponse       "Requested channel or ref does not exist"
// @Failure      404  {object}  libs.ErrResponse       "Arrow not found"
// @Failure      500  {object}  libs.ErrResponse       "Internal error"
// @Router       /arrow/{ns} [patch]
func (h *Handlers) Update(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))
	opts := models.UpdateOptions{}
	if c.Request.Body != nil {
		_ = c.ShouldBindJSON(&opts)
	}
	if _, err := h.svc.Update(c.Request.Context(), ns, opts); err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns), err)
		return
	}
	libs.WriteMutationOK(c, http.StatusOK, string(ns))
}

// Remove deregisters an arrow, addressed by the namespace it was registered
// under — bare (github.com/user/repo) or versioned (…@v1.0.0).
//
// @Summary      Remove arrow
// @Description  Deregisters an arrow, addressed by the namespace it was registered under (bare or versioned).
// @Tags         arrows
// @Param        ns   path  string  true  "Arrow namespace (bare or versioned, e.g. github.com/user/repo@v1.0.0)"
// @Success      200  {object}  libs.MutationResponse  "Arrow removed"
// @Failure      404  {object}  libs.ErrResponse       "Arrow not found"
// @Failure      500  {object}  libs.ErrResponse       "Internal error"
// @Router       /arrow/{ns} [delete]
func (h *Handlers) Remove(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))
	if err := h.svc.Remove(c.Request.Context(), ns); err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns), err)
		return
	}
	libs.WriteMutationOK(c, http.StatusOK, string(ns))
}

// List returns all registered arrows.
//
// @Summary      List arrows
// @Description  Returns all registered arrows. Use the WebSocket upgrade to stream real-time updates instead.
// @Tags         arrows
// @Produce      json
// @Param        user_installed  query  boolean  false  "Filter by user-installed flag"
// @Success      200  {object}  libs.QueryResponse{data=[]apidto.ArrowListItemDTO}
// @Failure      500  {object}  libs.ErrResponse
// @Router       /arrow [get]
func (h *Handlers) List(c *gin.Context) {
	var userInstalled *bool
	if param := c.Query("user_installed"); param != "" {
		v := param == "true"
		userInstalled = &v
	}
	items, err := h.svc.List(c.Request.Context(), userInstalled)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, "", err)
		return
	}
	dtos := make([]apidto.ArrowListItemDTO, 0, len(items))
	for _, item := range items {
		dtos = append(dtos, apidto.ArrowListItemDTOFrom(item))
	}
	libs.WriteQueryOK(c, dtos)
}

// GetDetail returns the full detail view of a single arrow including runtime state.
//
// @Summary      Get arrow detail
// @Description  Returns detailed information for an arrow including current state, active run, and last return. An uncatalogued namespace resolves live, the same way the manifest and readme endpoints do, and reports state "absent". Use the WebSocket upgrade to stream live updates.
// @Tags         arrows
// @Produce      json
// @Param        ns   path  string  true  "Arrow namespace"
// @Success      200  {object}  libs.QueryResponse{data=apidto.ArrowDetailDTO}
// @Failure      404  {object}  libs.ErrResponse
// @Failure      422  {object}  libs.ErrResponse
// @Failure      502  {object}  libs.ErrResponse
// @Failure      500  {object}  libs.ErrResponse
// @Router       /arrow/{ns} [get]
func (h *Handlers) GetDetail(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))
	detail, err := h.svc.GetDetail(c.Request.Context(), ns)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns), err)
		return
	}
	libs.WriteQueryOK(c, apidto.ArrowDetailDTOFrom(detail))
}

// @Summary      Preview arrow
// @Description  Resolves an arrow's manifest without registering it, the same way GetDetail resolves an uncatalogued namespace live. ?format=raw returns the manifest bytes instead of the parsed detail.
// @Tags         arrows
// @Produce      json
// @Param        ns      path   string  true   "Arrow namespace"
// @Param        format  query  string  false  "raw returns the manifest bytes as text/markdown"
// @Success      200  {object}  libs.QueryResponse{data=apidto.ArrowDetailDTO}
// @Failure      404  {object}  libs.ErrResponse
// @Failure      422  {object}  libs.ErrResponse
// @Failure      502  {object}  libs.ErrResponse
// @Failure      500  {object}  libs.ErrResponse
// @Router       /arrow/{ns}/preview [get]
func (h *Handlers) Preview(
	c *gin.Context,
) {
	ns := domain.Namespace(c.Param("ns"))
	detail, raw, err := h.svc.Preview(c.Request.Context(), ns)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns), err)
		return
	}
	if c.Query("format") == "raw" {
		c.Header("X-Content-Type-Options", "nosniff")
		c.Data(http.StatusOK, "text/markdown; charset=utf-8", raw)
		return
	}
	libs.WriteQueryOK(c, apidto.ArrowDetailDTOFrom(detail))
}

// GetManifest returns the raw manifest definition for an arrow.
//
// @Summary      Get arrow manifest
// @Description  Returns the full manifest definition including targets, variables, and lifecycle steps.
// @Tags         arrows
// @Produce      json
// @Param        ns   path  string  true  "Arrow namespace"
// @Success      200  {object}  libs.QueryResponse{data=apidto.ArrowManifestDTO}
// @Failure      404  {object}  libs.ErrResponse
// @Failure      422  {object}  libs.ErrResponse
// @Failure      502  {object}  libs.ErrResponse
// @Failure      500  {object}  libs.ErrResponse
// @Router       /arrow/{ns}/manifest [get]
func (h *Handlers) GetManifest(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))
	result, err := h.svc.GetManifest(c.Request.Context(), ns)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns), err)
		return
	}
	libs.WriteQueryOK(c, apidto.ArrowManifestDTOFrom(result))
}

// GetReadme returns the prose documentation for an arrow.
//
// @Summary      Get arrow readme
// @Description  Returns the prose surrounding the fenced manifest block when the arrow is delivered as ARROW.md. 404 if the arrow was delivered as arrow.yaml or carries no readme.
// @Tags         arrows
// @Produce      json
// @Param        ns   path  string  true  "Arrow namespace"
// @Success      200  {object}  libs.QueryResponse{data=apidto.ArrowReadmeDTO}
// @Failure      404  {object}  libs.ErrResponse
// @Failure      422  {object}  libs.ErrResponse
// @Failure      502  {object}  libs.ErrResponse
// @Failure      500  {object}  libs.ErrResponse
// @Router       /arrow/{ns}/readme [get]
func (h *Handlers) GetReadme(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))
	result, err := h.svc.GetReadme(c.Request.Context(), ns)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns))
		return
	}
	libs.WriteQueryOK(c, apidto.ArrowReadmeDTOFrom(ns, result))
}

// GetDependents returns the arrows that depend on this arrow.
//
// @Summary      Get arrow dependents
// @Description  Returns the namespaces of arrows that declare a dependency on this arrow, across all installed versions pointing at it.
// @Tags         arrows
// @Produce      json
// @Param        ns   path  string  true  "Arrow namespace"
// @Success      200  {object}  libs.QueryResponse{data=apidto.ArrowDependentsDTO}
// @Failure      404  {object}  libs.ErrResponse
// @Failure      500  {object}  libs.ErrResponse
// @Router       /arrow/{ns}/dependents [get]
func (h *Handlers) GetDependents(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))
	dependents, err := h.svc.GetDependents(c.Request.Context(), ns)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns))
		return
	}
	libs.WriteQueryOK(c, apidto.ArrowDependentsDTOFrom(ns, dependents))
}

// GetDependencies returns the resolved, topologically ordered dependency plan for this arrow.
//
// @Summary      Get arrow dependencies
// @Description  Returns the transitive dependency plan for this arrow, resolved from its manifest's tools and services, in dependency-first order.
// @Tags         arrows
// @Produce      json
// @Param        ns   path  string  true  "Arrow namespace"
// @Success      200  {object}  libs.QueryResponse{data=apidto.ArrowDependenciesDTO}
// @Failure      404  {object}  libs.ErrResponse
// @Failure      500  {object}  libs.ErrResponse
// @Router       /arrow/{ns}/dependencies [get]
func (h *Handlers) GetDependencies(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))
	plan, err := h.svc.GetDependencies(c.Request.Context(), ns)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns))
		return
	}
	libs.WriteQueryOK(c, apidto.ArrowDependenciesDTOFrom(ns, plan))
}

// ListChannels reports every release channel an arrow's repository publishes.
//
// @Summary      List arrow channels
// @Description  Buckets every tag the arrow's repository publishes into its release channel, plus the repository's default branch.
// @Tags         arrows
// @Produce      json
// @Param        ns   path  string  true  "Arrow namespace"
// @Success      200  {object}  libs.QueryResponse{data=apidto.ChannelListDTO}  "Channels"
// @Failure      404  {object}  libs.ErrResponse  "Arrow not found"
// @Failure      500  {object}  libs.ErrResponse  "Internal error"
// @Router       /arrow/{ns}/channels [get]
func (h *Handlers) ListChannels(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))
	channels, err := h.svc.ListChannels(c.Request.Context(), ns)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns), err)
		return
	}
	libs.WriteQueryOK(c, apidto.ChannelListDTOFrom(channels))
}

// Seed uploads a raw YAML manifest for an arrow and registers it immediately.
//
// @Summary      Seed arrow manifest
// @Description  Accepts a raw YAML manifest in the request body, stores it in the registry, and registers the arrow. Use this to publish a new arrow from a local manifest file.
// @Tags         arrows
// @Accept       application/x-yaml
// @Param        ns    path  string  true   "Arrow namespace"
// @Param        body  body  string  true   "Raw YAML manifest"
// @Success      201   {object}  libs.MutationResponse  "Manifest seeded and arrow registered"
// @Failure      400   {object}  libs.ErrResponse       "Failed to read body"
// @Failure      422   {object}  libs.ErrResponse       "Invalid manifest"
// @Failure      500   {object}  libs.ErrResponse       "Internal error"
// @Router       /arrow/{ns}/manifest [post]
func (h *Handlers) Seed(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))

	body, err := io.ReadAll(c.Request.Body)
	if err != nil {
		libs.WriteErr(c, http.StatusBadRequest, "failed to read body", string(ns))
		return
	}

	if err := h.svc.Seed(c.Request.Context(), ns, body); err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns), err)
		return
	}

	libs.WriteMutationOK(c, http.StatusCreated, string(ns))
}

// Validate checks a raw YAML manifest without registering it.
//
// @Summary      Validate arrow manifest
// @Description  Parses and validates a raw YAML manifest. Returns validation errors and platform compatibility without writing anything to the registry.
// @Tags         arrows
// @Accept       application/x-yaml
// @Produce      json
// @Param        ns    path  string  true  "Arrow namespace"
// @Param        body  body  string  true  "Raw YAML manifest"
// @Success      200   {object}  libs.QueryResponse{data=apidto.ValidationResultDTO}  "Manifest is valid"
// @Failure      400   {object}  libs.ErrResponse                                     "Failed to read body"
// @Failure      422   {object}  libs.QueryResponse{data=apidto.ValidationResultDTO}  "Manifest has validation errors"
// @Failure      500   {object}  libs.ErrResponse                                  "Internal error"
// @Router       /arrow/{ns}/manifest/validate [post]
func (h *Handlers) Validate(c *gin.Context) {
	ns := domain.Namespace(c.Param("ns"))

	body, err := io.ReadAll(c.Request.Body)
	if err != nil {
		libs.WriteErr(c, http.StatusBadRequest, "failed to read body", string(ns))
		return
	}

	result, err := h.svc.ValidateManifest(c.Request.Context(), body)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, string(ns), err)
		return
	}

	dto := apidto.ValidationResultDTOFrom(result)
	if result.Valid {
		libs.WriteQueryOK(c, dto)
	} else {
		libs.WriteQueryWithStatus(c, http.StatusUnprocessableEntity, dto)
	}
}